LayoutTests/platform/ios-simulator/editing/spelling/context-menu-suggestions-expected.txt

CONSOLE MESSAGE: line 32: TypeError: undefined is not an object (evaluating 'itemNamesWithSpellChecking.length')
This test checks that spelling suggestions in the context menu are not show when spellcheck="false". To test manually, right click on the text in the input box.


FAIL: Did not run test
